# backend/swagger.yaml
openapi: 3.0.3
info:
  title: RTX Trading Engine API
  description: |
    Complete API documentation for RTX Trading Engine.
    
    ## Overview
    - **B-Book Trading**: Internal execution with RTX balance (NOT broker)
    - **Admin Panel**: Deposit, withdraw, adjust balances
    - **Market Data**: Real-time prices and historical data
    - **Legacy APIs**: OANDA passthrough (deprecated)
    
    ## Authentication
    Most endpoints require JWT authentication via Bearer token.
  version: 3.0.0
  contact:
    name: RTX Support
    email: support@rtx.com
servers:
  - url: http://localhost:8080
    description: Development Server
...
paths:
...
  /ticks:
    get:
      tags: [Market Data]
      summary: Get historical tick data
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
            example: EURUSD
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Tick data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tick'
  /ohlc:
    get:
      tags: [Market Data]
      summary: Get OHLC candle data
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: timeframe
          in: query
          schema:
            type: string
            enum: [1m, 5m, 15m, 30m, 1h, 4h, 1d]
            default: 1m
        - name: limit
          in: query
          schema:
            type: integer
            default: 500
      responses:
        '200':
          description: OHLC data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OHLC'

  # ===== RISK CALCULATOR =====
...
components:
  schemas:
...
    Tick:
      type: object
      properties:
        symbol:
          type: string
        bid:
          type: number
        ask:
          type: number
        spread:
          type: number
        lp:
          type: string
        timestamp:
          type: string
          format: date-time
    OHLC:
      type: object
      properties:
        time:
          type: integer
          description: Unix timestamp
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT